% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted_split.R
\name{check_duplicates}
\alias{check_duplicates}
\title{Undo 0-1 scaling.}
\usage{
check_duplicates(subsetter, path_num)
}
\arguments{
\item{subsetter}{The subsetting matrix.}

\item{path_num}{Total number of pathways.}
}
\value{
\code{is_a_duplicate}.
}
\description{
Undo 0-1 scaling.
}
