% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_table.R
\name{check_table}
\alias{check_table}
\title{Check if table populations are theoretically distinguishable.}
\usage{
check_table(plusminus_table)
}
\arguments{
\item{plusminus_table}{Table indicated whether each group (row) is positive
(+1), negative (-1), or neutral / unknown (0) for each
variable (column).}
}
\value{
List:
\itemize{
\item pass: Logical, whether the table passed the check.
\item indistinct_pops: List of indistinguishable sub-tables.
}
}
\description{
If all splits were achieved, is there enough information in the user-provided
table to distinguish between all of the described populations.
}
