% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_variables.R
\name{combine_variables}
\alias{combine_variables}
\title{Combine a pair of variables to find a line at the given angle.}
\usage{
combine_variables(x, y, degrees, min_max_scaling = NULL)
}
\arguments{
\item{x}{x axis data vector.}

\item{y}{y axis data vector.}

\item{degrees}{The range of degrees to be searched over. Overrides \code{slope_type}.}

\item{min_max_scaling}{Vector giving the new minimum and new maximum.}
}
\value{
A numeric vector.
}
\description{
Combine a pair of variables to find a line at the given angle.
}
