% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_f1.R
\name{compute_max_f1}
\alias{compute_max_f1}
\title{Compute the maximum F1 for a single class with respect to a single cluster.}
\usage{
compute_max_f1(
  class_labels,
  cluster_labels,
  weighted = TRUE,
  hungarian = FALSE,
  return_option = c("vector", "value", "matrix"),
  no_match_class = NULL,
  no_match_cluster = NULL
)
}
\arguments{
\item{class_labels}{Vector of true class labels.}

\item{cluster_labels}{Vector of estimated cluster labels.}

\item{weighted}{Logical: whether the mean should be weighted with respect to
the size of the classes.}

\item{hungarian}{Logical: whether the Hungarian algorithm should be used to
obtain a one-to-one mapping of classes to clusters.}

\item{return_option}{Character: whether to return the full pairwise F1
matrix ("matrix"), the vector of maximum F1 scores for
each class ("vector", default), or the mean ("value").}

\item{no_match_class}{Classes to be excluded when matching. Their F1 values
are set to -Inf.}

\item{no_match_cluster}{Clusters to be excluded when matching. Their F1
values are set to -Inf.}
}
\value{
Either a matrix, vector, or mean value, depending on \code{return_option}.
}
\description{
Compute a matrix of F1 scores for each class-cluster pair, then optionally
obtain maximum F1 values for each class, either independently or according
to a one-to-one mapping by the Hungarian algorithm. Finally, a weighted or
unweighted mean may be computed across the classes.
}
