% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_f1.R
\name{compute_recall_entropy}
\alias{compute_recall_entropy}
\title{Compute the entropy for a single class across all clusters.}
\usage{
compute_recall_entropy(
  class_labels,
  cluster_labels,
  return_option = c("vector", "value", "matrix"),
  weighted = TRUE
)
}
\arguments{
\item{class_labels}{Vector of true class labels.}

\item{cluster_labels}{Vector of estimated cluster labels.}

\item{return_option}{Character: whether to return the full pairwise F1
matrix ("matrix"), the vector of maximum F1 scores for
each class ("vector", default), or the mean ("value").}

\item{weighted}{Logical: whether the mean should be weighted with respect to
the size of the classes.}
}
\value{
Either a matrix, vector, or mean value, depending on \code{return_option}.
}
\description{
Compute a matrix of entropy scores for each class-cluster pair, then
optionally obtain the summed total entropy values for each class,
Finally, a weighted or unweighted mean may be computed across the classes.
}
