% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_table.R
\name{expand_table}
\alias{expand_table}
\title{Add missing variables to a plusminus table.}
\usage{
expand_table(x, plusminus_subtable)
}
\arguments{
\item{x}{Dataset in matrix or data.frame form.}

\item{plusminus_subtable}{A \code{matrix} or \code{data.frame} whose column names are a
subset of the column names of \code{x}.}
}
\value{
A \code{data.frame} with the same number of columns as \verb{x``.}
}
\description{
Given a plusminus subtable which does not include all of the variables from
the data set, add in zero columns for the unused variables.
}
