% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_split.R
\name{exploratory_plot}
\alias{exploratory_plot}
\title{Plotting function for exploratory_split.}
\usage{
exploratory_plot(
  x_gp,
  splits,
  g,
  p_choice,
  subsetter,
  scores,
  is_negative,
  var_name
)
}
\arguments{
\item{x_gp}{Univariate data.}

\item{splits}{split matrix.}

\item{g}{The current pathway.}

\item{p_choice}{The chosen variable.}

\item{subsetter}{The subsetting matrix.}

\item{scores}{The scores matrix.}

\item{is_negative}{Whether to display the negative or positive side.}

\item{var_name}{The name of the variable.}
}
\value{
A ggplot object.
}
\description{
Plotting function for exploratory_split.
}
