% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_labels.R
\name{extract_labels}
\alias{extract_labels}
\title{Extract labels from \code{gatetree} output.}
\usage{
extract_labels(gatetree_out)
}
\arguments{
\item{gatetree_out}{Output from \code{gatetree} function.}
}
\value{
List of label vectors for each sample.
}
\description{
Extract labels from \code{gatetree} output.
}
