% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_boundary_functions.R
\name{find_boundary}
\alias{find_boundary}
\title{Find the optimal univariate GMM boundary.}
\usage{
find_boundary(x, min_scaled_bic_diff = 0)
}
\arguments{
\item{x}{Dataset in \code{matrix} or \code{data.frame} form.}

\item{min_scaled_bic_diff}{Minimum value of difference between one-component
and two-component BIC divided by
2 log(obs_num).}
}
\value{
Vector consisting of the boundary and its scaled BIC difference. The
boundary will be NA if its scaled BIC difference is less than
\code{min_scaled_bic_diff}.
}
\description{
Find the optimal two-component univariate GMM boundary and check if its
scaled BIC difference is greater than \code{min_scaled_bic_diff}.
}
