% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peak_edge.R
\name{find_peak_edge}
\alias{find_peak_edge}
\title{Find the border between a Gaussian component a Uniform noise component.}
\usage{
find_peak_edge(x, after_peak = TRUE)
}
\arguments{
\item{x}{Data vector}

\item{after_peak}{Whether the algorithm should look for an edge to the right
of the peak.}
}
\value{
A scalar: the edge location.
}
\description{
Find the border between a Gaussian component a Uniform noise component.
}
