% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_valley.R
\name{find_peaks}
\alias{find_peaks}
\title{Find local maxima of a univariate kernel density estimate.}
\usage{
find_peaks(dens, width_percent = 1, min_height = 0.01)
}
\arguments{
\item{dens}{Output from \code{stats::density} function.}

\item{width_percent}{Radius of local window as a percentage of the range.}

\item{min_height}{Minimum height, as a percentage of the height of the global
density maximum, for a peak to be recognised by find_peaks.}
}
\value{
Logical vector indicating which points in \code{dens$x} are local maxima.
}
\description{
Find points whose kernel density is greater than the points in a
surrounding region. The size of this region is controlled by \code{width_percent}.
}
