% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_valley.R
\name{find_valley}
\alias{find_valley}
\title{Find the deepest univariate density valley.}
\usage{
find_valley(x, min_kde_size = 100)
}
\arguments{
\item{x}{Data vector}

\item{min_kde_size}{Minimum number of events required to search for a KDE valley.}
}
\value{
Vector of length two: valley location and depth
}
\description{
Find the deepest univariate density valley.
}
