% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_split_2d.R
\name{find_valley_2d}
\alias{find_valley_2d}
\title{Choose the slope and intercept for a two-dimensional valley split.}
\usage{
find_valley_2d(x, y, slope_type = NULL, degrees = NULL)
}
\arguments{
\item{x}{x axis data vector.}

\item{y}{y axis data vector.}

\item{slope_type}{\code{"incline"}, \code{"vertical"}, or \code{"decline"}.}

\item{degrees}{The range of degrees to be searched over. Overrides \code{slope_type}.}
}
\value{
A 1-row \code{data.frame} containing:
\describe{
\item{\code{split}}{Valley split value.}
\item{\code{score}}{Valley depth.}
\item{\code{intercept}}{Intercept for two-dimensional line.}
\item{\code{slope}}{Slope for two-dimensional line.}
\item{\code{alpha}}{Mixing coefficient for x-axis variable.}
\item{\code{beta}}{Mixing coefficient for y-axis variable.}
\item{\code{degree}}{Degree of chosen two-dimensional line.}
}
}
\description{
Choose the slope and intercept for a two-dimensional valley split.
}
