% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateTree-package.R
\docType{package}
\name{gateTree-package}
\alias{gateTree}
\alias{gateTree-package}
\title{gateTree: A user-informed, semi-supervised clustering tree}
\description{
A user-informed, semi-supervised clustering tree.
}
\author{
\strong{Maintainer}: Ultán P. Doherty \email{dohertyu@tcd.ie} (\href{https://orcid.org/0000-0003-1791-395X}{ORCID})

Authors:
\itemize{
  \item Arthur White (\href{https://orcid.org/0000-0002-7268-5163}{ORCID})
}

}
\keyword{internal}
