% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatetree.R
\name{gatetree}
\alias{gatetree}
\title{User-informed clustering decision tree.}
\usage{
gatetree(
  x,
  plusminus_table = expand.grid(rep(list(c(-1, 1)), ncol(x))),
  order_table = array(0, dim = dim(plusminus_table)),
  min_height = min_depth,
  min_depth = 1,
  min_scaled_bic_diff = 0,
  min_size = 50,
  min_val_cutoff = NULL,
  max_val_cutoff = NULL,
  temp_min_cutoff = NULL,
  temp_max_cutoff = NULL,
  use_boundaries = TRUE,
  show_plot = c(FALSE, FALSE),
  explore = TRUE
)
}
\arguments{
\item{x}{Dataset in \code{matrix} or \code{data.frame} form.}

\item{plusminus_table}{Table indicated whether each group (row) is positive
(+1), negative (-1), or neutral / unknown (0) for each
variable (column).}

\item{order_table}{Table influencing the order of splits.}

\item{min_height}{Minimum height, as a percentage of the height of the global
density maximum, for a peak to be recognised by
\link{find_peaks}.}

\item{min_depth}{Minimum depth, as a percentage of the height of the global
density maximum, for a split to be returned by
\link{find_valley}.}

\item{min_scaled_bic_diff}{Minimum value of difference between one-component
and two-component BIC divided by
2 log(obs_num).}

\item{min_size}{Minimum number of observations for a subset to be split.}

\item{min_val_cutoff}{Minimum values per variable for observations to be
included.}

\item{max_val_cutoff}{Maximum values per variable for observations to be
included.}

\item{temp_min_cutoff}{Minimum values per variable for observations to be
included in finding a split.}

\item{temp_max_cutoff}{Maximum values per variable for observations to be
included in finding a split.}

\item{use_boundaries}{Logical value.}

\item{show_plot}{Logical vector of length 2. Indicating whether the split
plots (\verb{[1]}) and the tree plot (\verb{[2]}) should be plotted.}

\item{explore}{Logical value.}
}
\value{
A \code{list} object:
\itemize{
\item \code{splits}: matrix of split locations for all variables and described
populations.
\item \code{split_order}: matrix of split order for all variables and described
populations.
\item \code{edge_df}: data.frame describing the tree's edges and nodes.
\item \code{labels}: vector of integer cluster labels. \code{0} represents "Unassigned".
\item \code{signs}: matrix of split signs for all variables and described populations.
\item \code{tree_plot}: the clustering tree as a \code{ggplot} object.
}
}
\description{
Construct a semi-supervised decision tree to identify user-described
clusters.
}
