% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_f1.R
\name{get_f1}
\alias{get_f1}
\title{Compute F1 given two vectors of labels}
\usage{
get_f1(
  clust_labels,
  class_labels,
  remove_class = NULL,
  no_match_class = NULL,
  prec_rec = FALSE
)
}
\arguments{
\item{clust_labels}{Vector of cluster labels to evaluate.}

\item{class_labels}{Vector of true / reference labels to compare to.}

\item{remove_class}{Vector of class labels to be removed from the data.}

\item{no_match_class}{Vector of class labels to which cluster labels should
not be matched.}

\item{prec_rec}{Logical: whether precision and recall vectors are returned.}
}
\value{
List: F1 matrix & F1 vector, and if prec_rec is TRUE, precision
vector and recall vector.
}
\description{
Compute F1 given two vectors of labels
}
