% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_plotting_functions.R
\name{make_edge_df}
\alias{make_edge_df}
\title{Compile edge and node information.}
\usage{
make_edge_df(
  parent_node,
  node_number,
  edge_name,
  node_name,
  is_leaf,
  path_nodes
)
}
\arguments{
\item{parent_node}{Integer vector: The number of each node's parent node.}

\item{node_number}{Integer: The total number of nodes.}

\item{edge_name}{Character vector: The name of each edge.}

\item{node_name}{Character vector: The name of each node.}

\item{is_leaf}{Logical vector: whether each node is a leaf / terminal node.}

\item{path_nodes}{List of nodes on each path.}
}
\value{
\code{data.frame} with columns:
\itemize{
\item parent
\item node
\item edge_name
\item node_name
\item is_leaf
\item ggraph_order
}
}
\description{
Compile edge and node information into a \code{data.frame} for \link{make_tree_plot}.
}
