% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_plotting_functions.R
\name{make_tree_plot}
\alias{make_tree_plot}
\title{Construct tree diagram for \code{gateTree}.}
\usage{
make_tree_plot(edge_df, show_plot = FALSE)
}
\arguments{
\item{edge_df}{Output from \link{make_edge_df}.}

\item{show_plot}{Logical: should the tree diagram be plotted?}
}
\value{
\code{ggraph} object.
}
\description{
Construct a tree diagram from the information contained in \code{edge_df}.
}
