% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_split.R
\name{plot_explore}
\alias{plot_explore}
\title{Density plot with a \code{gateTree} split.}
\usage{
plot_explore(samples, gatetree_out, pop = NULL, samp = NULL, var = NULL)
}
\arguments{
\item{samples}{Dataset in \code{matrix} or \code{data.frame} form.}

\item{gatetree_out}{Output from gatetree.}

\item{pop}{The population number.}

\item{samp}{The sample number.}

\item{var}{The variable number.}
}
\value{
\code{ggplot} object.
}
\description{
Plot a univariate density of the data with the \code{gateTree} split
illustrated, and information about the split in the title and subtitle.
}
