% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_gatetree_split}
\alias{plot_gatetree_split}
\title{Plotting function used in \link{gatetree}.}
\usage{
plot_gatetree_split(x_gp, g, p, score, plusminus_table, scenario, split_gp)
}
\arguments{
\item{x_gp}{The data to be displayed, should be only pathway g & variable p.}

\item{g}{The pathway number.}

\item{p}{The variable number.}

\item{score}{The valley depth percentage or the scaled BIC difference.}

\item{plusminus_table}{Table indicated whether each group (row) is positive
(+1), negative (-1), or neutral / unknown (0) for each
variable (column).}

\item{scenario}{\code{"valley"}, \code{"boundary"}, \code{"nothing"}, or \code{"explore"}.}

\item{split_gp}{The split value.}
}
\value{
\code{ggplot} object.
}
\description{
Plotting function used in \link{gatetree}.
}
