% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_split.R
\name{plot_split}
\alias{plot_split}
\title{Density plot with a \code{gateTree} split.}
\usage{
plot_split(
  matrices,
  gatetree_out,
  pop = NULL,
  batch = NULL,
  samp = NULL,
  var = NULL
)
}
\arguments{
\item{matrices}{Dataset in \code{matrix} or \code{data.frame} form.}

\item{gatetree_out}{Output from gatetree.}

\item{pop}{The population number.}

\item{batch}{The batch number.}

\item{samp}{The sample number.}

\item{var}{The variable number.}
}
\value{
\code{ggplot} object.
}
\description{
Plot a univariate density of the data with the \code{gateTree} split
illustrated, and information about the split in the title and subtitle.
}
