% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_split_overlay.R
\name{plot_split_overlay}
\alias{plot_split_overlay}
\title{Density plot with a \code{gateTree} split.}
\usage{
plot_split_overlay(matrices, gatetree_out, var, pop, batch = NULL, samp = NULL)
}
\arguments{
\item{matrices}{Dataset in \code{matrix} or \code{data.frame} form.}

\item{gatetree_out}{Output from gatetree.}

\item{var}{The variable number.}

\item{pop}{The population number.}

\item{batch}{The batch number.}

\item{samp}{The sample number.}
}
\value{
\code{ggplot} object.
}
\description{
Plot a univariate density of the data with the \code{gateTree} split
illustrated, and information about the split in the title and subtitle.
}
