% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{plot_tree}
\alias{plot_tree}
\title{gateTree2 tree plot.}
\usage{
plot_tree(gatetree_out)
}
\arguments{
\item{gatetree_out}{Output from gatetree.}
}
\value{
\code{ggplot} object.
}
\description{
gateTree2 tree plot.
}
