% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_splits.R
\name{propose_boundaries}
\alias{propose_boundaries}
\title{Wrapper for \code{find_boundary}.}
\usage{
propose_boundaries(
  x,
  g,
  var_num,
  subsetter,
  already_split,
  common_variables = NULL
)
}
\arguments{
\item{x}{Data.}

\item{g}{The pathway number.}

\item{var_num}{Total number of variables.}

\item{subsetter}{The subsetting matrix.}

\item{already_split}{The already_split matrix.}

\item{common_variables}{Matrix}
}
\value{
boundaries
}
\description{
Wrapper for \code{find_boundary}.
}
