% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_splits.R
\name{propose_valleys}
\alias{propose_valleys}
\title{Wrapper for \code{find_valley}.}
\usage{
propose_valleys(
  x,
  g,
  var_num,
  subsetter,
  already_split,
  min_depth,
  min_height,
  common_variables = NULL
)
}
\arguments{
\item{x}{Data.}

\item{g}{The pathway number.}

\item{var_num}{Total number of variables.}

\item{subsetter}{The subsetting matrix.}

\item{already_split}{The already_split matrix.}

\item{min_depth}{\code{min_depth} for \code{find_valley} function.}

\item{min_height}{\code{min_height} for \code{find_valley} function.}

\item{common_variables}{Matrix}
}
\value{
valleys
}
\description{
Wrapper for \code{find_valley}.
}
