% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_splits.R
\name{propose_valleys2}
\alias{propose_valleys2}
\title{Wrapper for \code{find_valley}.}
\usage{
propose_valleys2(
  x,
  subsetter_g,
  splittable_vars_g = rep(TRUE, ncol(x)),
  min_depth,
  min_height
)
}
\arguments{
\item{x}{Data.}

\item{subsetter_g}{Column g of the subsetting matrix.}

\item{splittable_vars_g}{Row g of the splittable_vars matrix.}

\item{min_depth}{\code{min_depth} for \code{find_valley} function.}

\item{min_height}{\code{min_height} for \code{find_valley} function.}
}
\value{
valleys
}
\description{
Wrapper for \code{find_valley}.
}
