% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_variables.R
\name{reconstruct_2d_line}
\alias{reconstruct_2d_line}
\title{Compute the slope and intercept of a two-dimensional line for a univariate
split from a combined variable.}
\usage{
reconstruct_2d_line(split, x, y, degrees, min_max_scaling = NULL)
}
\arguments{
\item{split}{Value of the split for the combined variable.}

\item{x}{x axis data vector.}

\item{y}{y axis data vector.}

\item{degrees}{The range of degrees to be searched over. Overrides \code{slope_type}.}

\item{min_max_scaling}{Vector giving the new minimum and new maximum.}
}
\value{
A numeric vector.
}
\description{
Compute the slope and intercept of a two-dimensional line for a univariate
split from a combined variable.
}
