% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted_branches.R
\name{targeted_branches}
\alias{targeted_branches}
\title{Partition data set based on marginal splits described by +/- table.}
\usage{
targeted_branches(
  x,
  plusminus_table,
  min_height = 0.5,
  min_depth = 0.5,
  min_val_cutoff = NULL,
  max_val_cutoff = NULL,
  use_boundaries = TRUE,
  explore = TRUE
)
}
\arguments{
\item{x}{Dataset in matrix or data.frame form.}

\item{plusminus_table}{Table indicated whether each group (row) is positive
(+1), negative (-1), or neutral / unknown (0) for each
variable (column).}

\item{min_height}{Minimum height, as a percentage of the height of the global
density maximum, for a peak to be recognised by find_peaks.}

\item{min_depth}{Minimum depth, as a percentage of the height of the global
density maximum, for a split to be returned by find_valley.}

\item{min_val_cutoff}{Minimum value for an observation to be included.}

\item{max_val_cutoff}{Maximum value for an observation to be included.}

\item{use_boundaries}{Logical value.}

\item{explore}{Logical value.}
}
\value{
splits, plusminus_table, subsetter, plot_list
}
\description{
Partition data set based on marginal splits described by +/- table.
}
