% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted_split.R
\name{targeted_split}
\alias{targeted_split}
\title{Partition data set based on marginal splits described by +/- table.}
\usage{
targeted_split(
  x,
  typemarker,
  min_height = 0.5,
  min_depth = 0.5,
  min_val_cutoff = NULL,
  max_val_cutoff = NULL,
  use_boundaries = TRUE,
  explore = TRUE
)
}
\arguments{
\item{x}{Dataset in matrix or data.frame form.}

\item{typemarker}{Cell-type marker table.}

\item{min_height}{Minimum height for a peak to be recognised by find_peaks.}

\item{min_depth}{Minimum depth for a split to be returned by find_valley.}

\item{min_val_cutoff}{Minimum value for an observation to be included.}

\item{max_val_cutoff}{Maximum value for an observation to be included.}

\item{use_boundaries}{Logical value.}

\item{explore}{Logical value.}
}
\value{
splits, typemarker, subsetter
}
\description{
Partition data set based on marginal splits described by +/- table.
}
