% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted_tree.R
\name{targeted_tree}
\alias{targeted_tree}
\title{Partition data set based on unsupervised marginal splits.}
\usage{
targeted_tree(
  x,
  min_height = 0.5,
  min_depth = 0.5,
  min_val_cutoff = NULL,
  max_val_cutoff = NULL,
  show_plot = FALSE,
  plusminus_table,
  use_boundaries = TRUE
)
}
\arguments{
\item{x}{Dataset in matrix or data.frame form.}

\item{min_height}{Minimum height, as a percentage of the height of the global
density maximum, for a peak to be recognised by find_peaks.}

\item{min_depth}{Minimum depth, as a percentage of the height of the global
density maximum, for a split to be returned by find_valley.}

\item{min_val_cutoff}{Minimum value for an observation to be included.}

\item{max_val_cutoff}{Maximum value for an observation to be included.}

\item{show_plot}{Logical value.}

\item{plusminus_table}{Table indicated whether each group (row) is positive
(+1), negative (-1), or neutral / unknown (0) for each
variable (column).}

\item{use_boundaries}{Logical value.}
}
\value{
splits, typemarker, subsetter
}
\description{
Partition data set based on unsupervised marginal splits.
}
