% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted_split.R
\name{unscale01}
\alias{unscale01}
\title{Undo 0-1 scaling.}
\usage{
unscale01(x, unscaled_min, unscaled_max)
}
\arguments{
\item{x}{Data.}

\item{unscaled_min}{Minimum used for 0-1 scaling.}

\item{unscaled_max}{Maximum used for 0-1 scaling.}
}
\value{
Unscaled version of \code{x}.
}
\description{
Undo 0-1 scaling.
}
