% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_auxiliary_functions.R
\name{check_duplicates}
\alias{check_duplicates}
\title{Check if final subsets are identical.}
\usage{
check_duplicates(subsetter)
}
\arguments{
\item{subsetter}{The subsetting matrix.}
}
\value{
Vector indicating whether a cluster contains the same observations as
an existing cluster. If a pair of clusters are identical, the second one is
identified as a duplicate.
}
\description{
Check if final subsets are identical.
}
