% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auxiliary.R
\name{check_duplicates}
\alias{check_duplicates}
\title{Check if final subsets are identical.}
\usage{
check_duplicates(subsetter)
}
\arguments{
\item{subsetter}{The subsetting matrix.}
}
\value{
\code{is_a_duplicate}.
}
\description{
Check if final subsets are identical.
}
