% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_f1.R
\name{compute_merged_f1}
\alias{compute_merged_f1}
\title{Compute the F1 for each class of the optimal merged set of cluster.}
\usage{
compute_merged_f1(class_labels, cluster_labels, merge_limit = NULL)
}
\arguments{
\item{class_labels}{Vector of true class labels.}

\item{cluster_labels}{Vector of estimated cluster labels.}

\item{merge_limit}{Maximum number of cluster merges to perform per class.}
}
\value{
\code{compute_merged_f1} returns a list with the following elements:
\describe{
\item{\code{f1}}{Vector: max merged F1 values for each class.}
\item{\code{merges}}{Vector: number of merges performed for each class.}
}
}
\description{
For each class, merge clusters in order from highest to lowest precision,
stopping when the F1 of the merged cluster with respect to that class is
maximised.
}
