% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_boundary.R
\name{find_boundary}
\alias{find_boundary}
\title{Title}
\usage{
find_boundary(x, min_scaled_bic_diff = 0, return_all = FALSE)
}
\arguments{
\item{x}{Dataset in matrix or data.frame form.}

\item{min_scaled_bic_diff}{Minimum value of difference between one-component
and two-component BIC divided by
2 log(obs_num).}

\item{return_all}{}
}
\value{
to be filled in
}
\description{
Title
}
