% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_boundary.R
\name{find_boundary}
\alias{find_boundary}
\title{Find the optimal univariate GMM boundary.}
\usage{
find_boundary(x, boundary_noise_comp = TRUE)
}
\arguments{
\item{x}{Data vector}

\item{boundary_noise_comp}{= TRUE}
}
\value{
Vector consisting of the boundary and its scaled BIC difference. The
boundary will be NA if its scaled BIC difference is less than
\code{min_scaled_bic_diff}.
}
\description{
Find the optimal two-component univariate GMM boundary and check if its
scaled BIC difference is greater than \code{min_scaled_bic_diff}.
}
