% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_auxiliary_functions.R
\name{find_inside_cutoffs}
\alias{find_inside_cutoffs}
\title{Find which observations are inside the cutoffs for each variable.}
\usage{
find_inside_cutoffs(x, min_val_cutoff, max_val_cutoff)
}
\arguments{
\item{x}{Matrix}

\item{min_val_cutoff}{Minimum value vector}

\item{max_val_cutoff}{Maximum value vector}
}
\value{
Matrix of the same dimensions as \code{x} containing logical values
indicating whether that observation (row) is inside the cutoffs for that
variable (column).
}
\description{
Find which observations are inside the cutoffs for each variable.
}
