% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted_split.R
\name{find_inside_cutoffs}
\alias{find_inside_cutoffs}
\title{Find which observations are inside the cutoffs for each marker.}
\usage{
find_inside_cutoffs(x, min_val_cutoff, max_val_cutoff)
}
\arguments{
\item{x}{Dataset in matrix or data.frame form.}

\item{min_val_cutoff}{Minimum value for an observation to be included.}

\item{max_val_cutoff}{Maximum value for an observation to be included.}
}
\value{
inside_cutoffs
}
\description{
Find which observations are inside the cutoffs for each marker.
}
