% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_valley.R
\name{find_valley}
\alias{find_valley}
\title{Title}
\usage{
find_valley(
  dens,
  is_peak = NULL,
  return_depth = FALSE,
  min_depth = 0.01,
  min_height = 0.01
)
}
\arguments{
\item{dens}{Output from \code{stats::density} function.}

\item{is_peak}{Logical vector for peaks.}

\item{return_depth}{Logical: should valley depth be outputted?}

\item{min_depth}{Minimum depth, as a percentage of the height of the global
density maximum, for a split to be returned by find_valley.}

\item{min_height}{Minimum height, as a percentage of the height of the global
density maximum, for a peak to be recognised by find_peaks.}
}
\value{
Valley location or valley location and valley depth.
}
\description{
Title
}
