% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_valley_functions.R
\name{find_valley}
\alias{find_valley}
\title{Find the optimal density valley.}
\usage{
find_valley(dens, min_height = 0.01)
}
\arguments{
\item{dens}{Output from \code{stats::density} function.}

\item{min_height}{Minimum peak height when searching for density valleys.}
}
\value{
Vector consisting of the valley and its depth percentage. The
valley will be NA if its depth percentage is less than \code{min_depth}.
}
\description{
Find the optimal kernel density valley and check if its depth percentage is
greater than \code{min_depth}.
}
