% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_valley.R
\name{find_valley}
\alias{find_valley}
\title{Find the deepest univariate density valley.}
\usage{
find_valley(x)
}
\arguments{
\item{x}{Data vector}
}
\value{
Vector of length two: valley location and depth
}
\description{
Find the deepest univariate density valley.
}
