% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatetree.R
\name{gatetree}
\alias{gatetree}
\title{User-informed clustering decision tree.}
\usage{
gatetree(
  samples,
  plusminus,
  min_depth = 100,
  min_diff = 0.05,
  use_gmm = TRUE,
  min_cutoffs = NULL,
  max_cutoffs = NULL,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{samples}{Dataset in \code{matrix} or \code{data.frame} form.}

\item{plusminus}{Table indicating whether each population (row) is positive (+1),
negative (-1), or neutral / unknown (0) for each variable (column).}

\item{min_depth}{Minimum depth, as a percentage of the height of the global density maximum,
for a split to be returned by \link{find_valley}.}

\item{min_diff}{Minimum value of difference between one-component and two-component BIC
divided by 2 log(obs_num).}

\item{use_gmm}{Logical value.}

\item{min_cutoffs}{Minimum values for observations used when finding splits.}

\item{max_cutoffs}{Maximum values for observations used when finding splits.}

\item{seed}{Random seed for GMM fitting.}

\item{verbose}{Logical value.}
}
\value{
A \code{list} object:
}
\description{
Construct a semi-supervised decision tree to identify user-described
clusters.
}
\examples{

iris_plusminus <- rbind(
  "setosa" = c(0, 0, -1, 0),
  "versicolor" = c(0, 0, +1, -1),
  "virginica" = c(0, 0, +1, +1)
)
colnames(iris_plusminus) <- colnames(iris_plusminus)[1:4]

iris_gatetree <- gatetree(
  list(iris[, -5]),
  iris_plusminus,
  min_depth = 10,
  min_diff = 0.01
)
}
