% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatetree_classify.R
\name{gatetree_classify}
\alias{gatetree_classify}
\title{Classify a data set based on splits outputted by the \code{gatetree} function.}
\usage{
gatetree_classify(x, gatetree_output)
}
\arguments{
\item{x}{Dataset in matrix or data.frame form.}

\item{gatetree_output}{Output from the \code{gatetree} function.}
}
\value{
Vector of integer labels.
}
\description{
Classify a data set based on splits outputted by the \code{gatetree} function.
}
