% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_experimental.R
\name{merge_subsets}
\alias{merge_subsets}
\title{Supervised merging.}
\usage{
merge_subsets(signs, typemarker, subsetter)
}
\arguments{
\item{signs}{The signs matrix.}

\item{typemarker}{The cell type - marker matrix.}

\item{subsetter}{The subsetting matrix.}
}
\value{
A ggplot object.
}
\description{
Supervised merging.
}
