% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted_split.R
\name{plot_targeted_split}
\alias{plot_targeted_split}
\title{Plotting function for \code{targeted_split}.}
\usage{
plot_targeted_split(
  x_gp,
  g,
  p,
  depth,
  typemarker,
  scenario,
  split_gp,
  split_size
)
}
\arguments{
\item{x_gp}{The data to be displayed, should be only pathway g and marker p.}

\item{g}{The pathway number.}

\item{p}{The marker number.}

\item{depth}{The depth of the split.}

\item{typemarker}{The cell-type marker table.}

\item{scenario}{"valley", "boundary", "nothing", or "undiscovered".}

\item{split_gp}{The split value.}

\item{split_size}{The size of the subset before and after the split.}
}
\description{
Plotting function for \code{targeted_split}.
}
