% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_splits.R
\name{propose_boundaries}
\alias{propose_boundaries}
\title{Wrapper for \code{find_valley}.}
\usage{
propose_boundaries(x, subsetter_g, splittable_vars_g = rep(TRUE, ncol(x)))
}
\arguments{
\item{x}{Data.}

\item{subsetter_g}{Column g of the subsetting matrix.}

\item{splittable_vars_g}{Row g of the splittable_vars matrix.}
}
\value{
valleys
}
\description{
Wrapper for \code{find_valley}.
}
