% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted_split.R
\name{propose_boundaries}
\alias{propose_boundaries}
\title{Wrapper for \code{find_boundary}.}
\usage{
propose_boundaries(x, g, var_num, subsetter, progress)
}
\arguments{
\item{x}{Data.}

\item{g}{The pathway number.}

\item{var_num}{Total number of markers.}

\item{subsetter}{The subsetting matrix.}

\item{progress}{The progress matrix.}
}
\value{
boundaries
}
\description{
Wrapper for \code{find_boundary}.
}
