% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_boundary.R
\name{propose_boundaries}
\alias{propose_boundaries}
\title{Wrapper for \code{find_boundary}.}
\usage{
propose_boundaries(
  x,
  min_scaled_bic_diff = 0,
  subsetter_g,
  splittable_vars_g = rep(TRUE, ncol(x))
)
}
\arguments{
\item{x}{Dataset in matrix or data.frame form.}

\item{min_scaled_bic_diff}{Minimum value of difference between one-component
and two-component BIC divided by
2 log(obs_num).}

\item{subsetter_g}{Column g of the subsetting matrix.}

\item{splittable_vars_g}{Row g of the splittable_vars matrix.}
}
\value{
boundaries
}
\description{
Wrapper for \code{find_boundary}.
}
