% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_boundary_functions.R
\name{propose_boundaries}
\alias{propose_boundaries}
\title{Apply \link{find_boundary} across a set of variables.}
\usage{
propose_boundaries(
  x,
  min_scaled_bic_diff = 0,
  subsetter_g,
  splittable_vars_g = rep(TRUE, ncol(x))
)
}
\arguments{
\item{x}{Dataset in matrix or data.frame form.}

\item{min_scaled_bic_diff}{Minimum value of difference between one-component
and two-component BIC divided by
2 log(obs_num).}

\item{subsetter_g}{Column g of the subsetting matrix.}

\item{splittable_vars_g}{Row g of the splittable_vars matrix.}
}
\value{
Matrix in which the columns contain each variable's boundary and its
scaled BIC difference.
}
\description{
Apply \link{find_boundary} to each variable in a set of splittable
variables for a subset of observations.
}
