% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_splits.R
\name{propose_splits}
\alias{propose_splits}
\title{Propose a density valley or GMM boundary.}
\usage{
propose_splits(
  x,
  subsetter_g,
  splittable_vars_g,
  min_size,
  min_depth,
  min_height = min_depth,
  min_scaled_bic_diff = 0,
  use_boundaries
)
}
\arguments{
\item{x}{Dataset in \code{matrix} or \code{data.frame} form.}

\item{subsetter_g}{Column g of the subsetting matrix.}

\item{splittable_vars_g}{Row g of the splittable_vars matrix.}

\item{min_size}{Minimum number of observations for a subset to be split.}

\item{min_depth}{Minimum depth, as a percentage of the height of the global
density maximum, for a split to be returned by
\link{find_valley}.}

\item{min_height}{Minimum height, as a percentage of the height of the global
density maximum, for a peak to be recognised by
\link{find_peaks}.}

\item{min_scaled_bic_diff}{Minimum value of difference between one-component
and two-component BIC divided by
2 log(obs_num).}

\item{use_boundaries}{Logical value.}
}
\value{
List
\itemize{
\item matrix: each column contains the split location for that variable and the
valley depth percentage or scaled BIC difference of that split.
\item scenario: a character string indicating the type of the split (\code{"valley"}
or \code{"boundary"}), if no split was found (\code{"nothing"}), or if the split was
identified in the context of the \code{explore} option (\code{"explore"}).
}
}
\description{
Find a density valley or GMM boundary given a subset of cells and a set of
splittable variables.
}
