% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted_split.R
\name{propose_valleys}
\alias{propose_valleys}
\title{Wrapper for \code{find_valley}.}
\usage{
propose_valleys(x, g, var_num, subsetter, progress, min_depth, min_height)
}
\arguments{
\item{x}{Data.}

\item{g}{The pathway number.}

\item{var_num}{Total number of markers.}

\item{subsetter}{The subsetting matrix.}

\item{progress}{The progress matrix.}

\item{min_depth}{\code{min_depth} for \code{find_valley} function.}

\item{min_height}{\code{min_height} for \code{find_valley} function.}
}
\value{
valleys
}
\description{
Wrapper for \code{find_valley}.
}
