% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_valley.R
\name{propose_valleys}
\alias{propose_valleys}
\title{Wrapper for \code{find_valley}.}
\usage{
propose_valleys(
  x,
  subsetter_g,
  splittable_vars_g = rep(TRUE, ncol(x)),
  min_depth,
  min_height
)
}
\arguments{
\item{x}{Dataset in matrix or data.frame form.}

\item{subsetter_g}{Column g of the subsetting matrix.}

\item{splittable_vars_g}{Row g of the splittable_vars matrix.}

\item{min_depth}{Minimum depth, as a percentage of the height of the global
density maximum, for a split to be returned by find_valley.}

\item{min_height}{Minimum height, as a percentage of the height of the global
density maximum, for a peak to be recognised by find_peaks.}
}
\value{
valleys
}
\description{
Wrapper for \code{find_valley}.
}
