% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_auxiliary.R
\name{scale01}
\alias{scale01}
\title{0-1 scaling.}
\usage{
scale01(x, other_min = NULL, other_max = NULL)
}
\arguments{
\item{x}{Dataset in matrix or data.frame form.}

\item{other_min}{Minimum to be used for 0-1 scaling.}

\item{other_max}{Maximum to be used for 0-1 scaling.}
}
\value{
Scaled version of \code{x}.
}
\description{
0-1 scaling.
}
