% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted_tree.R
\name{targeted_tree}
\alias{targeted_tree}
\title{Partition data set based on unsupervised marginal splits.}
\usage{
targeted_tree(
  x,
  plusminus_table = expand.grid(rep(list(c(-1, 1)), nrow(x))),
  min_height = min_depth,
  min_depth = 1,
  min_val_cutoff = NULL,
  max_val_cutoff = NULL,
  use_boundaries = TRUE,
  show_plot = FALSE
)
}
\arguments{
\item{x}{Dataset in matrix or data.frame form.}

\item{plusminus_table}{Table indicated whether each group (row) is positive
(+1), negative (-1), or neutral / unknown (0) for each
variable (column).}

\item{min_height}{Minimum height, as a percentage of the height of the global
density maximum, for a peak to be recognised by find_peaks.}

\item{min_depth}{Minimum depth, as a percentage of the height of the global
density maximum, for a split to be returned by find_valley.}

\item{min_val_cutoff}{Minimum value for an observation to be included.}

\item{max_val_cutoff}{Maximum value for an observation to be included.}

\item{use_boundaries}{Logical value.}

\item{show_plot}{Logical value.}
}
\value{
List: splits, split_order, subsetter, edge_df, labels, signs.
}
\description{
Partition data set based on unsupervised marginal splits.
}
